+++ .env.example
@@ -1,5 +1,6 @@
-# Nutrient DWS API Configuration
+# Nutrient DWS Processor API Configuration for Testing
 NUTRIENT_API_KEY=your_api_key_here
+NUTRIENT_BASE_URL=https://api.nutrient.io
 
 # Development Settings
 DEBUG=true
