+++ .github/workflows/ci.yml
@@ -50,7 +50,7 @@ jobs:
       run: npm ci
     
     - name: Run unit tests with coverage
-      run: npm test -- --coverage --testPathPatterns='^((?!integration|e2e).)*$'
+      run: npm test -- --coverage --testPathPatterns='^((?!integration).)*$'
     
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
@@ -86,77 +86,3 @@ jobs:
         test -f dist/index.js
         test -f dist/index.cjs
         test -f dist/index.d.ts
-    
-    - name: Upload build artifacts
-      uses: actions/upload-artifact@v4
-      with:
-        name: build-artifacts
-        path: dist/
-
-  release:
-    runs-on: ubuntu-latest
-    needs: [build]
-    if: github.ref == 'refs/heads/main' && github.event_name == 'push'
-    permissions:
-      contents: write
-      packages: write
-    
-    steps:
-    - uses: actions/checkout@v4
-      with:
-        fetch-depth: 0
-    
-    - name: Use Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '20.x'
-        cache: 'npm'
-        registry-url: 'https://registry.npmjs.org'
-    
-    - name: Install dependencies
-      run: npm ci
-    
-    - name: Download build artifacts
-      uses: actions/download-artifact@v4
-      with:
-        name: build-artifacts
-        path: dist/
-    
-    - name: Check if version changed
-      id: version-check
-      run: |
-        CURRENT_VERSION=$(node -p "require('./package.json').version")
-        git fetch --tags
-        if git rev-list -n 1 "v$CURRENT_VERSION" 2>/dev/null; then
-          echo "Version $CURRENT_VERSION already released"
-          echo "should_release=false" >> $GITHUB_OUTPUT
-        else
-          echo "New version $CURRENT_VERSION detected"
-          echo "should_release=true" >> $GITHUB_OUTPUT
-          echo "version=$CURRENT_VERSION" >> $GITHUB_OUTPUT
-        fi
-    
-    - name: Create Release
-      if: steps.version-check.outputs.should_release == 'true'
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      run: |
-        gh release create "v${{ steps.version-check.outputs.version }}" \
-          --title "Release v${{ steps.version-check.outputs.version }}" \
-          --notes "Automated release for version ${{ steps.version-check.outputs.version }}" \
-          --target ${{ github.sha }}
-    
-    - name: Publish to NPM
-      if: steps.version-check.outputs.should_release == 'true'
-      run: |
-        if [ -n "$NODE_AUTH_TOKEN" ]; then
-          echo "Setting up NPM authentication..."
-          echo "//registry.npmjs.org/:_authToken=${NODE_AUTH_TOKEN}" > ~/.npmrc
-          echo "Publishing to NPM..."
-          npm publish --access public
-        else
-          echo "⚠️ Skipping NPM publish - NPM_TOKEN not configured"
-          echo "To enable NPM publishing, add NPM_TOKEN to repository secrets"
-        fi
-      env:
-        NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
