+++ .github/workflows/integration-tests.yml
@@ -1,4 +1,4 @@
-name: E2E Tests
+name: Integration Tests
 
 on:
   push:
@@ -7,7 +7,7 @@ on:
     branches: [ main ]
 
 jobs:
-  e2e-tests:
+  integration-tests:
     runs-on: ubuntu-latest
     needs: []  # Run in parallel with other workflows
     
@@ -37,17 +37,17 @@ jobs:
           echo "✅ Valid API key detected"
         else
           echo "has_api_key=false" >> $GITHUB_OUTPUT
-          echo "⏭️ No valid API key - E2E tests will be skipped"
+          echo "⏭️ No valid API key - Integration tests will be skipped"
         fi
     
-    - name: Run E2E tests
+    - name: Run integration tests
       if: steps.check-api-key.outputs.has_api_key == 'true'
       env:
         NUTRIENT_API_KEY: ${{ secrets.NUTRIENT_API_KEY }}
-      run: npm run test:e2e
+      run: npm run integration
     
-    - name: Skip E2E tests (no API key)
+    - name: Skip integration tests (no API key)
       if: steps.check-api-key.outputs.has_api_key == 'false'
       run: |
-        echo "✅ E2E tests skipped - no valid API key available"
+        echo "✅ Integration tests skipped - no valid API key available"
         echo "This is expected for forks and external PRs"
