+++ .github/workflows/scheduled-integration-tests.yml
@@ -1,4 +1,4 @@
-name: Scheduled E2E Tests
+name: Scheduled Integration Tests
 
 on:
   schedule:
@@ -7,7 +7,7 @@ on:
   workflow_dispatch:  # Allow manual triggering
 
 jobs:
-  scheduled-e2e-tests:
+  scheduled-integration-tests:
     runs-on: ubuntu-latest
     
     steps:
@@ -22,13 +22,13 @@ jobs:
     - name: Install dependencies
       run: npm ci
     
-    - name: Run all E2E tests
+    - name: Run all integration tests
       if: secrets.NUTRIENT_API_KEY != ''
       env:
         NUTRIENT_API_KEY: ${{ secrets.NUTRIENT_API_KEY }}
       run: |
-        echo "Running scheduled E2E tests to detect API changes..."
-        npm test -- --testPathPatterns="(integration|e2e-advanced)\.test\.ts$" --verbose --passWithNoTests
+        echo "Running scheduled integration tests to detect API changes..."
+        npm test -- --testPathPatterns="integration\.test\.ts$" --verbose --passWithNoTests
       timeout-minutes: 20
       continue-on-error: true
       id: test-run
@@ -36,16 +36,16 @@ jobs:
     - name: Skip scheduled tests (no API key)
       if: secrets.NUTRIENT_API_KEY == ''
       run: |
-        echo "⏭️ Skipping scheduled E2E tests - NUTRIENT_API_KEY not available"
+        echo "⏭️ Skipping scheduled integration tests - NUTRIENT_API_KEY not available"
         echo "Configure NUTRIENT_API_KEY secret to enable scheduled API validation"
     
     - name: Generate detailed test report
       if: always()
       run: |
-        npm test -- --json --outputFile=scheduled-test-results.json --testPathPatterns="(integration|e2e-advanced)\.test\.ts$" --passWithNoTests || true
+        npm test -- --json --outputFile=scheduled-test-results.json --testPathPatterns="integration\.test\.ts$" --passWithNoTests || true
         
         # Create summary
-        echo "## E2E Test Summary" > test-summary.md
+        echo "## Integration Test Summary" > test-summary.md
         echo "Date: $(date)" >> test-summary.md
         echo "Status: ${{ steps.test-run.outcome }}" >> test-summary.md
         
@@ -67,13 +67,13 @@ jobs:
       with:
         script: |
           const date = new Date().toISOString().split('T')[0];
-          const title = `🚨 E2E Tests Failed - ${date}`;
+          const title = `🚨 Integration Tests Failed - ${date}`;
           
           // Check if issue already exists
           const issues = await github.rest.issues.listForRepo({
             owner: context.repo.owner,
             repo: context.repo.repo,
-            labels: ['e2e-failure', 'automated'],
+            labels: ['integration-failure', 'automated'],
             state: 'open'
           });
           
@@ -84,9 +84,9 @@ jobs:
               owner: context.repo.owner,
               repo: context.repo.repo,
               title: title,
-              body: `## Scheduled E2E tests failed
+              body: `## Scheduled Integration tests failed
               
-              The scheduled E2E test run has detected failures. This could indicate:
+              The scheduled integration test run has detected failures. This could indicate:
               - API changes that need to be addressed
               - Service degradation
               - Test flakiness
@@ -101,7 +101,7 @@ jobs:
               See the workflow artifacts for detailed test results.
               
               cc @jdrhyne`,
-              labels: ['e2e-failure', 'automated', 'high-priority']
+              labels: ['integration-failure', 'automated', 'high-priority']
             });
           }
     
@@ -109,7 +109,7 @@ jobs:
       uses: actions/upload-artifact@v4
       if: always()
       with:
-        name: scheduled-e2e-results-${{ github.run_number }}
+        name: scheduled-integration-results-${{ github.run_number }}
         path: |
           scheduled-test-results.json
           test-summary.md
@@ -120,11 +120,11 @@ jobs:
       uses: actions/github-script@v7
       with:
         script: |
-          // Close any open E2E failure issues
+          // Close any open integration failure issues
           const issues = await github.rest.issues.listForRepo({
             owner: context.repo.owner,
             repo: context.repo.repo,
-            labels: ['e2e-failure', 'automated'],
+            labels: ['integration-failure', 'automated'],
             state: 'open'
           });
           
@@ -134,6 +134,6 @@ jobs:
               repo: context.repo.repo,
               issue_number: issue.number,
               state: 'closed',
-              body: issue.body + '\n\n---\n\n✅ **Resolved**: E2E tests are now passing.'
+              body: issue.body + '\n\n---\n\n✅ **Resolved**: Integration tests are now passing.'
             });
           }
