+++ examples/.env.example
@@ -1,2 +1,2 @@
-# Nutrient DWS API Key
+# Nutrient DWS Processor API Key
 NUTRIENT_API_KEY=your_api_key_here
