+++ package.json
@@ -76,9 +76,8 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
-    "test:unit": "jest --testPathPatterns=\"^((?!integration|e2e).)*$\"",
+    "test:unit": "jest --testPathPatterns=\"^((?!integration).)*$\"",
     "test:integration": "jest --testPathPatterns=integration.test",
-    "test:e2e": "jest --testPathPatterns=\"(integration|e2e-advanced).test\"",
     "test:integration:verbose": "jest --testPathPatterns=integration.test --verbose",
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix",
